requirements.txt
pdpyras==5.1.2
pytz==2023.3.post1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
tabulate==0.9.0
